fix bugs elastic search

+++ src/services/sequelize.service.js
@@ -13,28 +13,32 @@ const modelFiles = fs
     .readdirSync(__dirname + "/../models/")
     .filter((file) => file.endsWith(".js"));
 
+let sequelize;
 const sequelizeService = {
     init: async () => {
         try {
-            const sequelize = new Sequelize(databaseConfig);
+            if (!sequelize) {
+                sequelize = new Sequelize(databaseConfig);
 
-            /*
-              Loading models automatically
-            */
+                /*
+                  Loading models automatically
+                */
 
-            for (const file of modelFiles) {
-                const model = await import(`../models/${file}`);
-                model.default.init(sequelize);
+                for (const file of modelFiles) {
+                    const model = await import(`../models/${file}`);
+                    model.default.init(sequelize);
+                }
+
+                modelFiles.map(async (file) => {
+                    const model = await import(`../models/${file}`);
+                    model.default.associate && model.default.associate(sequelize.models);
+                });
+
+                sequelize.sync({ force: false }).then(res => {
+                    console.log("[SEQUELIZE] Database service initialized");
+                })
             }
-
-            modelFiles.map(async (file) => {
-                const model = await import(`../models/${file}`);
-                model.default.associate && model.default.associate(sequelize.models);
-            });
-
-            sequelize.sync({ force: false }).then(res => {
-                console.log("[SEQUELIZE] Database service initialized");
-            })
+            return sequelize;
 
         } catch (error) {
             console.log("[SEQUELIZE] Error during database service initialization");
